+++ packages/element/__tests__/decorators/attribute.ts
@@ -1,0 +1,4 @@
+const testAttribute = () => {
+  describe("@attribute", () => {
+  });
+};
