+++ packages/utils/__tests__/index.ts
@@ -1,11 +1,9 @@
 import assertKindTest from "./assertKind";
 import getSuperMethodTest from "./getSuperMethod";
-import testRegistry from "./Registry";
 import useInitializerTest from "./useInitializer";
 
 describe("@corpuscule/utils", () => {
   assertKindTest();
   useInitializerTest();
   getSuperMethodTest();
-  testRegistry();
 });
