+++ scripts/project.js
@@ -1,18 +1,18 @@
 const packages = {
-  context: ["index"],
-  element: [
-    "computed",
-    "dhtml",
-    "index",
-  ],
-  redux: ["index"],
-  router: ["index"],
-  styles: ["index"],
+  // context: ["index"],
+  // element: [
+  //   "computed",
+  //   "dhtml",
+  //   "index",
+  // ],
+  // redux: ["index"],
+  // router: ["index"],
+  // styles: ["index"],
   utils: [
-    "addToRegistry",
     "assertKind",
-    "getSuperMethod",
     "compose",
+    "getSuperMethod",
+    "useInitializer",
   ],
 };
